GAM-25 Added profile validation
