+++ packages/client/src/pages/EditProfilePage/EditProfilePage.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import {TextFieldProfile} from '../../components/ui';
 import {Avatar, Button} from '@mui/material';
 import {Form, Formik} from 'formik';
-import {SIGNUP_VALIDATION_SCHEMA} from "../SignupPage/validation-schema";
-import {INITIAL_FORM_STATE} from "../../services/initial-profile-states"
+import {EDIT_PROFILE_VALIDATION_SCHEMA, INITIAL_FORM_STATE} from "./validation-schema"
 
 export const EditProfilePage = (): JSX.Element => {
     const handleSubmit = () => console.log('Форма отправлена..')
@@ -12,7 +11,7 @@ export const EditProfilePage = (): JSX.Element => {
         <div className="profile">
             <Formik
                 initialValues={{...INITIAL_FORM_STATE}}
-                validationSchema={SIGNUP_VALIDATION_SCHEMA}
+                validationSchema={EDIT_PROFILE_VALIDATION_SCHEMA}
                 onSubmit={handleSubmit}
             >
                 <Form>
