--- packages/client/src/services/initial-profile-states.ts
@@ -1,11 +1,0 @@
-export const INITIAL_FORM_STATE = {
-  email: 'test@yandex.ru',
-  login: 'test_login',
-  first_name: 'First_name',
-  second_name: 'Second_name',
-  display_name: 'Display_name',
-  phone: '89991234567',
-  old_password: '',
-  password: '',
-  password_again: '',
-};
